Create 1897. Redistribute Characters to Make All Strings Equal Leetcode
You are given an array of strings words (0-indexed).

In one operation, pick two distinct indices i and j, where words[i] is a non-empty string, and move any character from words[i] to any position in words[j].

Return true if you can make every string in words equal using any number of operations, and false otherwise.

 

Example 1:

Input: words = ["abc","aabc","bc"]
Output: true
Explanation: Move the first 'a' in words[1] to the front of words[2],
to make words[1] = "abc" and words[2] = "abc".
All the strings are now equal to "abc", so return true.


ntuition

The intuition behind this problem is to check if it is possible to distribute the characters of all the words in the given vector in such a way that each character appears an equal number of times in all words.

Approach

The approach in the provided code is to use an unordered map (mpp) to store the count of each character in all words. After counting, the code checks if the count of each character is divisible by the total number of words (length). If the count of any character is not divisible evenly, it means it's not possible to distribute the characters equally among all words, and the function returns false. Otherwise, it returns true.

Complexity

Time complexity:
Time complexity: O(n * m), where n is the nstrumber of words and m is the average length of the words. The nested loops iterate through each character in each word.

Space complexity:
Space complexity: O(c), where c is the number of unique characters across all words. The unordered map stores the count of each unique character.